Follow dependency graph
After checking a particular resource, trigger checks on dependent
resources according to the dependency graph. This means we will create
all resources in dependency order. Yay!

To handle the case where a resource is dependent on multiple ancestors,
introduce a datastore purely for counting the dependencies satisfied.
This reduces the lock contention to the minimum possible domain.