Store dependency graph in Stack rather than resources
Storing the dependency graph centrally turns out to require much less
database traffic than storing each resource's dependencies locally
inside the resource. We would ultimately have needed to add dependencies
in both directions so that traversals in the reverse direction (for
cleanups and stack deletes) would be possible, and that would have meant
multiple passes through the database just to store all of the resources.

It has been suggested in https://review.openstack.org/#/c/123749/ that
we can avoid the problem of possibly having to rewrite the dependencies
during an update depending on whether a resource needs to be replaced or
not by _always_ creating new Resource entries in the database for every
resource in the updated template at the start of an update. It's unclear
at this stage if that will cause other issues. In any event, we would
still have had that problem when storing the dependencies in the
resources, due to the need to store the edges in both directions.