Implement update with replacement (and caveats!)
There's an important bug here: when we replace a resource, we have no
way of knowing whether to update the graph so that resources depending
on the previous one should now depend on the new one. If the dependent
resources are replaced, we don't want their dependencies redirected. If
they are then we do, although not yet - we should wait until the update
actually happens, since if it does not go ahead then the graph will be
left in the wrong state. None of this will affect the stack update, but
it will result in the incorrect cleanup dependencies being generated on
subsequent updates, so a failure during this update will likely prevent
the correct cleanups occurring later. I haven't figured out the exact
pathology.

There's also the issue that this is going to require us to do some sort
of locking on the stack in order to update the dependency graph. That's
something we really, really don't want to have to do.