Pass cleanup dependencies through messages
Take advantage of the fact that the initial stack load will have all of
the resource objects in memory and the fact that the cleanup graph
doesn't change over an update to calculate the cleanup dependencies at
the beginning (in much the same way that we calculate the update
dependencies from the template). This avoids having to load each target
resource to determine what its other requirements are, which is highly
inefficient in terms of database access.