Mark the stack as complete when all resources are updated
Right now this is implemented to move to the complete status when all of
the resources requested by the user are in their updated state, even if
there are other resources remaining to be cleaned up. This may or may
not be what we want (it certainly differs from the current behaviour of
Heat), but should be easy to change depending on what we decide.

For efficiency we only wait on resources that are root nodes in the
stack (i.e. nothing depends on them). However, if we wanted to cache
output values in the Stack column of the DB then we could change it to
aggregate the input_data from every resource (or we could get really,
really clever and only pass the needed data from resources that supply
it).