Use a unique ID to identify traversals
Previously we used the template ID to identify the traversal. However
this is not necessarily unique, as the user can roll back to a previous
template. That could result in operations from a previous, interrupted
attempt to update continuing while a second traversal also starts. That
could have advantages, but most likely its a bit too clever and will
result in painful debugging later on.

This patch provides an incrementing traversal ID instead.