Persist dependencies in Stack table
Rather than passing the dependencies down the chain as a parameter to
the check notifications, store them in the Stack table of the DB since
this will be loaded by all resources anyway.

We may be able to use this data later to restart interrupted traversals.
The biggest downside is that it requires an extra write to the Stack,
which is kind of a pain although I expect the performance impact would
be negligible.