Delete previous SyncPoints on update
This allows us to indicate to any resources still in progress from a
previous traversal that when the current action is complete they should
trigger an examination of the corresponding node for the new update.

Note, however, that this cannot happen in the context of this test
framework, so any handling code would never be executed.