Remove list of predecessor resources from SyncPoints
We already store two copies of the dependency graph - one distributed
through the Resource table and another for the current traversal in the
Stack table - so there is no need for a third to be stored in the
SyncPoints. SyncPoints only exist for the latest traversal, so when we
require the graph we can always obtain it from the Stack table.