Update resource requirements in two steps
Since we now update the Resource table twice for each resource update
(first to move to IN_PROGRESS and then to COMPLETE, as Heat itself
does), also update the dependencies in two steps: first adding the new
dependencies and then removing any that are out of date.

This ensures that if the update of a resource is interrupted (either by
the update failing, or by the engine dying) we have the maximal set of
possible dependencies stored to ensure the correct delete order is
enforced.

Note that this is not well tested at present, because there are no tests
for updates being interrupted.