Eliminate check for replaced resource
When resource update with replacement was introduced (by
77ccf686bdfcd8b47caf9196e1f570b3e8f37da5), it was necessary to avoid
updating a resource that had already been replaced. However, the very
next patch (ee48d93a5f4b14ec29ff835d0838abaa3cfadde5) eliminated the
need to do that by generating the traversal graph at the start of the
stack update (so we'll never end up checking the wrong resource).

In practice this check has resulted in a number of bugs in Heat, such as
https://storyboard.openstack.org/#!/story/2001974 - in part due to the
fact that the prototype does not simulate failed resources. The check is
also an obstacle to other enhancements, such as
https://storyboard.openstack.org/#!/story/2003579

For now, add an assertion error to verify that this case is not in fact
getting hit in any of the current scenarios. If future test scenarios
depend on this check _not_ being present (this would depend on some kind
of modelling of failed resources) then the assertion could be removed
altogether.